+++ organization/src/epilot/organization.py
@@ -61,7 +61,7 @@ class Organization:
         
         url = utils.generate_url(operations.UpdateOrganizationRequest, base_url, '/v2/organization/{org_id}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "organization", False, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.UpdateOrganizationRequest, "organization", False, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         headers['Accept'] = 'application/json'