+++ organization/src/epilot/organization_settings.py
@@ -101,7 +101,7 @@ class OrganizationSettings:
         
         url = utils.generate_url(operations.PutSettingsValueRequest, base_url, '/v2/organization/{org_id}/settings/{key}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "settings_value", False, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.PutSettingsValueRequest, "settings_value", False, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         headers['Accept'] = 'application/json'